Rename LED subsystem for purpose and allow for multiple LED subsystems
